feat: 작가 작품 관리 페이지 컴포넌트화 함
