+++ components/layouts/webtoonlistfooter/WebtoonListFooter.module.css
@@ -1,0 +1,29 @@
+footer.AuthorFooterWrap {
+  width: 90%;
+  height: 100px;
+  /* background: linear-gradient(180deg, #ffffff00 0%, #ffffff 50%); */
+  position: fixed;
+  bottom: 0rem;
+  z-index: 100;
+  display: flex;
+  justify-content: center;
+}
+
+.footerBtn {
+  width: 80px;
+  height: 80px;
+  background-color: var(--bp-primary);
+  border-radius: 50%;
+  box-shadow: 0px 5px 5px 2px #69d2ff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  opacity: 0.9;
+  filter: contrast(1);
+}
+
+.footerBtn img {
+  width: 30%;
+  height: 30%;
+  filter: brightness(0) invert(1);
+}
