+++ components/pages/authorworkslist/AuthorListCategory.module.css
@@ -1,0 +1,24 @@
+.HistoryCategory {
+  width: 100%;
+}
+
+.HistoryCategory nav ul {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.HistoryCategory nav ul li {
+  list-style: none;
+  width: 50%;
+  height: 2rem;
+  text-align: center;
+  font-size: 1rem;
+  font-weight: 500;
+}
+
+.HistoryCategory nav ul li.active {
+  font-weight: 600;
+  color: var(--bp-blue);
+  border-bottom: 0.2rem solid var(--bp-blue);
+}
